+++ app/controllers/PostsController.php
@@ -3,7 +3,6 @@
 class PostsController extends \BaseController
 {
 
-
     protected $layout = 'layouts.default';
 
     /**
@@ -14,8 +13,13 @@ class PostsController extends \BaseController
      */
     public function index()
     {
-        $this->layout->content = View::make('posts/index')
+        if (Auth::check()) {
+        $this->layout->content = View::make('posts.index')
             ->with('posts', Post::all());
+        } else {
+        $this->layout->content = View::make('guest.posts.index')
+            ->with('posts', Post::all());
+        }
     }
 
     /**
@@ -26,8 +30,14 @@ class PostsController extends \BaseController
      */
     public function create()
     {
-        //Render Create View
-        $this->layout->content = View::make('posts.create');
+        if (Auth::check()) {
+            //Render Create View
+            $this->layout->content = View::make('posts.create');
+        } else {
+            return Redirect::to('/')
+                ->with('flash_message', 'Solo el administrador puede crear nuevos articulos')
+                ->with('flash_type', 'alert-danger');
+        }
     }
 
     /**
@@ -72,10 +82,17 @@ class PostsController extends \BaseController
     {
         //Error handler
         try {
+            if (Auth::check()) {
             //Instantiate the record to show
             $post = Post::findOrFail($id);
             //Render show page with the record data
             $this->layout->content = View::make('posts.show', compact('post'));
+            } else {
+                //Instantiate the record to show
+                $post = Post::findOrFail($id);
+                //Render show page with the record data
+                $this->layout->content = View::make('guest.posts.show', compact('post'));
+            }
         } catch (\Exception $e) {
             return Redirect::to('posts')
                 ->with('flash_message', 'Algo salio mal. Error: ' . $e->getMessage())
@@ -94,10 +111,14 @@ class PostsController extends \BaseController
     {
         //Error handler
         try {
-            //Instantiate the record to show
-            $post = Post::findOrFail($id);
-            //Render show page with the record data
-            $this->layout->content = View::make('posts.edit', compact('post'));
+            if (Auth::check()) {
+                //Instantiate the record to show
+                $post = Post::findOrFail($id);
+                //Render show page with the record data
+                $this->layout->content = View::make('posts.edit', compact('post'));
+            } else {
+                throw new Exception('El Post solo puede ser modificado por el administrador');
+            }
         } catch (\Exception $e) {
             return Redirect::to('posts')
                 ->with('flash_message', 'Algo salio mal. Error: ' . $e->getMessage())
@@ -145,18 +166,22 @@ class PostsController extends \BaseController
     public function destroy($id)
     {
         try {
-            //Instantiate the record to edit
-            $posts = Post::findOrFail($id);
+            if (Auth::check()) {
+                //Instantiate the record to edit
+                $posts = Post::findOrFail($id);
 
-            //Delete the record from the DB
-            Post::find($id)->delete();
+                //Delete the record from the DB
+                Post::find($id)->delete();
 
-            //Check register still exist in the DB
-            if (empty(Post::find($id))) {
-                //Redirect to the photo.index page
-                return Redirect::to('posts')
-                    ->with('flash_message', 'El Post "' . $posts->title . '" se ha eliminado correctamente')
-                    ->with('flash_type', 'alert-success');
+                //Check register still exist in the DB
+                if (empty(Post::find($id))) {
+                    //Redirect to the photo.index page
+                    return Redirect::to('posts')
+                        ->with('flash_message', 'El Post "' . $posts->title . '" se ha eliminado correctamente')
+                        ->with('flash_type', 'alert-success');
+                } else {
+                    throw new Exception('El Post solo puede ser eliminado por el administrador');
+                }
             }
             throw new Exception('El Post "' . $posts->title . '" no se puedo eliminar. Si el error continua, contacte con su administrador');
         } catch (\Exception $e) {
